+++ inst/include/uwot/smooth_knn.h
@@ -66,19 +66,25 @@ struct SmoothKnnWorker {
   double double_max = (std::numeric_limits<double>::max)();
 
   std::vector<double> nn_weights;
+  
+  bool save_sigmas;
+  std::vector<double> sigmas;
 
   std::atomic_size_t n_search_fails;
 
   SmoothKnnWorker(const std::vector<double> &nn_dist, std::size_t n_vertices,
                   std::size_t n_iter, double local_connectivity,
-                  double bandwidth, double tol, double min_k_dist_scale)
+                  double bandwidth, double tol, double min_k_dist_scale,
+                  bool save_sigmas)
       : nn_dist(nn_dist), n_vertices(n_vertices),
         n_neighbors(nn_dist.size() / n_vertices),
         target(std::log2(n_neighbors)), n_iter(n_iter),
         local_connectivity(local_connectivity), bandwidth(bandwidth), tol(tol),
         min_k_dist_scale(min_k_dist_scale),
         mean_distances(mean_average(nn_dist)),
-        nn_weights(n_vertices * n_neighbors), n_search_fails(0) {}
+        nn_weights(n_vertices * n_neighbors), save_sigmas(save_sigmas), 
+        sigmas(save_sigmas ? n_vertices : 0), 
+        n_search_fails(0) {}
 
   void operator()(std::size_t begin, std::size_t end) {
     // number of binary search failures in this window
@@ -182,6 +188,9 @@ struct SmoothKnnWorker {
       }
 
       uwot::set_row(nn_weights, n_vertices, n_neighbors, i, res);
+      if (save_sigmas) {
+        sigmas[i] = sigma;
+      }
     }
 
     // Update global count of failures