+++ tests/testthat/test_fuzzy_simplicial_set.R
@@ -7,28 +7,28 @@ context("fuzzy simplicial set")
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_union, tol = 1e-4)
 
 # mix union + intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0.5, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_mix, tol = 1e-4)
 
 # intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect, tol = 1e-4)
 
 # Union + local_connectivity
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1.5,
   bandwidth = 1, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_local, tol = 1e-4)
 
 
@@ -36,14 +36,14 @@ expect_equal(res, V_union_local, tol = 1e-4)
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 0.5, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_bandwidth, tol = 1e-4)
 
 # intersect + local + bandwidth
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1.5,
   bandwidth = 0.5, verbose = FALSE, n_threads = 0, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect_local_bandwidth, tol = 1e-4)
 
 
@@ -52,28 +52,28 @@ expect_equal(res, V_intersect_local_bandwidth, tol = 1e-4)
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union, tol = 1e-4)
 
 # mix union + intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0.5, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_mix, tol = 1e-4)
 
 # intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect, tol = 1e-4)
 
 # Union + local_connectivity
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1.5,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_local, tol = 1e-4)
 
 
@@ -81,14 +81,14 @@ expect_equal(res, V_union_local, tol = 1e-4)
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 0.5, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_bandwidth, tol = 1e-4)
 
 # intersect + local + bandwidth
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1.5,
   bandwidth = 0.5, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect_local_bandwidth, tol = 1e-4)
 
 # parallel code path
@@ -96,28 +96,28 @@ expect_equal(res, V_intersect_local_bandwidth, tol = 1e-4)
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union, tol = 1e-4)
 
 # mix union + intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0.5, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_mix, tol = 1e-4)
 
 # intersection
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect, tol = 1e-4)
 
 # Union + local_connectivity
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1.5,
   bandwidth = 1, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_local, tol = 1e-4)
 
 
@@ -125,12 +125,12 @@ expect_equal(res, V_union_local, tol = 1e-4)
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 1, local_connectivity = 1,
   bandwidth = 0.5, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_union_bandwidth, tol = 1e-4)
 
 # intersect + local + bandwidth
 res <- fuzzy_simplicial_set(
   set_op_mix_ratio = 0, local_connectivity = 1.5,
   bandwidth = 0.5, verbose = FALSE, n_threads = 1, nn = nn
-)
+)$matrix
 expect_equal(res, V_intersect_local_bandwidth, tol = 1e-4)